+++ build.clj
@@ -3,7 +3,7 @@
   (:require [clojure.tools.build.api :as b]))
 
 (def lib 'build/micro-x)
-(def version "v0.12.116")
+(def version "v0.13.134")
 (def main 'chat.server)
 (def class-dir "target/classes")